+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## UNRELEASED
 
-Support `verbose.ignore` and `verbose.(extra|ignore)[].regex`. (#4 and #5)
+-   Support `no_verbose` and `regex` in the verbose rules. (#4 and #5)
+-   Support `explicitOnly` in the verbose rules. (#6)
+-   Restructure the verbose rules.
 
 ## 0.3.2